Removed unnecessary conversions to lists
I think it was 2to3 that wrapped all occurrences of my_var.keys(),
my_var.values(), and my_var.items() in list().  As far as I can tell,
almost all of these conversions are unnecessary.  This revision
removes all conversions when used in iterator context (which is most
of the time) but leaves them in when used in list context.