Drop support for --discard and --all-solns in favor of --show
I've even confused myself with the semantics of --discard and
--all-solns and how those relate to -O, -v, and -v -v.  I believe that
the new --show={valid, best, all} option is a more intuitive
replacement.  By default, only valid solutions are shown, which might
be the empty set.  Specifying --show=best ensures that at least one
solution (the least bad) is output.  Specifying --show=all outputs all
solutions.

This commit includes substantial restructuring of the code that
manages the resulting solutions.  A new Solutions object provides
methods for filtering out solutions that violate particular
constraints.  This is much cleaner than the previous approach, which
performed some filtering in dwave.py and some in qmasm.py, with much
confusion as to which solutions were truly the ones to be presented to
the user.