:bug: Fix incorrect handling of QUBO problems and their resulting energy
Programs run with `--qubo` were being incorrectly converted to Ising
problems (QMASM's native problem format) when all of the linear
coefficients in the generated Ising formulation are supposed to be 0.

The reported energy values (for both QUBO and Ising problems, I
believe) were also incorrect due to the problem coefficients being
scaled for D-Wave execution but the energy not being un-scaled back
into the terms of the original problem.

Thanks to Tom Krauss (@twist-vector) for the bug report and for
testing the updated code.