More aggressively merge qubits
Previously, user-defined chains would be converted to aliases (i.e.,
multiple qubits would be merged) only when the qubits at each end of
the chain have the same weight.  This change aliases qubits regardless
of weights.  Mathematically, this ought to be a correct operation, and
so far, the change does not appear to have degraded correctness.